== Happy New Year == *Added script command checklang() for checking player's language.
*Players can now use items and self-targetting skills while talking to NPCs.

*SL_STIN
*SL_STUN
-These skills can now be casted by non players.

*Optimized pc_useitem processing a bit.

*New ranklist packets will now be used on newer clients.

*Some cleaning.

*Added int64 support to status_charge function.

+++ src/char_sql/inter.c
@@ -244,7 +244,7 @@ static int inter_config_read(const char* cfgName)
 		}
 #ifndef TXT_SQL_CONVERT
 		else if(!strcmpi(w1,"party_share_level"))
-			party_share_level = atoi(w2);
+			party_share_level = (unsigned int)atof(w2);
 		else if(!strcmpi(w1,"log_inter"))
 			log_inter = atoi(w2);
 		else if(!strcmpi(w1,"main_chat_nick"))
